# src/api.yaml
openapi: 3.0.3
info:
  title: HD1 (Holodeck One) API
  description: |
    Revolutionary API-first 3D coordinate system with universal world boundaries.
    Single source of truth for sessions, worlds, objects, and real-time control.
    
    **CORE PRINCIPLES:**
    - Everything API-driven (zero shell commands)
    - Universal coordinate system [-12, +12] on all axes
    - Objects honor world boundaries absolutely
    - Named object persistence with full lifecycle
    - Real-time WebSocket + REST hybrid
    
    **ARCHITECTURE:**
    - Specification drives code generation
    - Build fails if handlers missing
    - Zero manual routing - auto-generated from spec
    
  version: 1.0.0
  contact:
    name: Holodeck One 3D Engine
servers:
  - url: http://localhost:8080/api
    description: Development server
paths:
...
  /sessions/{sessionId}:
    get:
      operationId: getSession
      tags: [Sessions]
      summary: Get session details
      description: Retrieve detailed session information including all objects
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/sessions/get.go"
      x-function: "GetSession"
      
    delete:
      operationId: deleteSession
      tags: [Sessions]
      summary: Terminate session
      description: Permanently delete session and all associated objects
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session terminated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/sessions/delete.go"
      x-function: "DeleteSession"

  # ========================================
  # SESSION GRAPH SESSION GRAPH & ROOM MANAGEMENT CHANNEL MANAGEMENT
  # TCP-level reliability with channel-based architecture
  # ========================================
...
  /sessions/{sessionId}/entities/{entityId}/components:
    get:
      operationId: listEntityComponents
      tags: [Components]
      summary: List all components on entity
      description: |
        Returns all PlayCanvas components attached to the specified entity.
        Components define entity capabilities like rendering, physics, audio, etc.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
          description: Filter components by type
      responses:
        '200':
          description: Entity components list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  entity_id:
                    type: string
                    example: "entity-abc123"
                  components:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComponentSummary'
                  total:
                    type: integer
                    example: 3
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/components/list.go"
      x-function: "ListEntityComponents"

    post:
      operationId: addComponent
      tags: [Components]
      summary: Add component to entity
      description: |
        Adds a new PlayCanvas component to the specified entity.
        Components provide functionality like rendering, physics, audio, etc.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                  enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
                  description: Component type to add
                  example: "model"
                properties:
                  type: object
                  description: Component-specific configuration properties
                  example:
                    type: "box"
                    castShadows: true
                    receiveShadows: true
                enabled:
                  type: boolean
                  default: true
                  description: Whether component is initially enabled
      responses:
        '201':
          description: Component added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  entity_id:
                    type: string
                    example: "entity-abc123"
                  component_id:
                    type: string
                    example: "comp-model-xyz789"
                  type:
                    type: string
                    example: "model"
                  enabled:
                    type: boolean
                    example: true
                  created_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Component type already exists on entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-handler: "api/components/add.go"
      x-function: "AddComponent"
  /sessions/{sessionId}/entities/{entityId}/components/{componentType}:
    get:
      operationId: getComponent
      tags: [Components]
      summary: Get component details
      description: |
        Retrieves detailed information about a specific component,
        including all properties, configuration, and current state.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
        - name: componentType
          in: path
          required: true
          schema:
            type: string
            enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
          example: "model"
      responses:
        '200':
          description: Component details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComponentDetail'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/components/get.go"
      x-function: "GetComponent"

    put:
      operationId: updateComponent
      tags: [Components]
      summary: Update component properties
      description: |
        Updates component properties and configuration.
        Only specified properties are modified, others remain unchanged.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
        - name: componentType
          in: path
          required: true
          schema:
            type: string
            enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
          example: "model"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                properties:
                  type: object
                  description: Component properties to update
                  example:
                    castShadows: false
                    material: "material-asset-id"
                enabled:
                  type: boolean
                  description: Enable/disable component
      responses:
        '200':
          description: Component updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  entity_id:
                    type: string
                    example: "entity-abc123"
                  component_type:
                    type: string
                    example: "model"
                  updated_properties:
                    type: array
                    items:
                      type: string
                    example: ["castShadows", "material"]
                  updated_at:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/components/update.go"
      x-function: "UpdateComponent"

    delete:
      operationId: removeComponent
      tags: [Components]
      summary: Remove component from entity
      description: |
        Permanently removes a component from the entity.
        All component data and functionality will be lost.
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            pattern: '^session_[a-f0-9]{8}$'
          example: "session_a1b2c3d4"
        - name: entityId
          in: path
          required: true
          schema:
            type: string
          example: "entity-abc123"
        - name: componentType
          in: path
          required: true
          schema:
            type: string
            enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
          example: "model"
      responses:
        '204':
          description: Component removed successfully
        '404':
          $ref: '#/components/responses/NotFound'
      x-handler: "api/components/remove.go"
      x-function: "RemoveComponent"
...
components:
  schemas:
...
    SessionSummary:
      type: object
      required: [id, created, world_initialized, object_count, status]
      properties:
        id:
          type: string
        created:
          type: string
          format: date-time
        world_initialized:
          type: boolean
        object_count:
          type: integer
          minimum: 0
        status:
          type: string
          enum: [active, inactive, error]
    SessionDetail:
      allOf:
        - $ref: '#/components/schemas/SessionSummary'
        - type: object
          properties:
            objects:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/ObjectDetail'

    # Environment Schemas
...
    ComponentSummary:
      type: object
      required: [component_id, type, enabled]
      properties:
        component_id:
          type: string
          description: HD1 component identifier
          example: "comp-model-xyz789"
        type:
          type: string
          enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
          description: PlayCanvas component type
          example: "model"
        enabled:
          type: boolean
          description: Whether component is active
          example: true
        playcanvas_id:
          type: string
          description: PlayCanvas internal component ID
          example: "pc-comp-guid-123"
        properties_count:
          type: integer
          description: Number of configured properties
          example: 8
        last_updated:
          type: string
          format: date-time
    ComponentDetail:
      allOf:
        - $ref: '#/components/schemas/ComponentSummary'
        - type: object
          properties:
            properties:
              type: object
              description: All component properties and values
              example:
                type: "box"
                castShadows: true
                receiveShadows: true
                material: "material-asset-id"
            schema:
              type: object
              description: Component property schema and validation rules
              example:
                type:
                  type: "string"
                  enum: ["asset", "box", "sphere"]
                  default: "box"
                castShadows:
                  type: "boolean"
                  default: true
            events:
              type: array
              items:
                $ref: '#/components/schemas/ComponentEvent'
              description: Recent component events
            dependencies:
              type: array
              items:
                type: string
              description: Other component types this depends on
              example: ["camera", "light"]
    ComponentEvent:
      type: object
      required: [event_id, component_type, event_type, timestamp]
      properties:
        event_id:
          type: string
          example: "evt-comp-abc123"
        component_type:
          type: string
          enum: [model, camera, light, rigidbody, script, sound, animation, collision, element, particlesystem, render, sprite]
          example: "model"
        event_type:
          type: string
          enum: [property_changed, enabled, disabled, asset_loaded, error]
          example: "property_changed"
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          description: Event-specific details
          example:
            property: "castShadows"
            old_value: true
            new_value: false
        message:
          type: string
          description: Human-readable event description
          example: "Model component shadow casting disabled"
...
    Success:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          type: string
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...